+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added IPv4 validation rule
+- Added IPv6 validation rule
+- Added Mac Address validation rule
+- Added Luxembourg VAT number validation
+- Added German VAT number validation
+- Added Belgium VAT number validation
+- Added French VAT number validation
+- Added UK Postal Code validation
 
 ### Changed