+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - EAN8 validation rule
 - EAN5 validation rule
 - EAN13 validation rule
+- UPC-A validation rule
 
 ### Changed