+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - South Korean (ROK) postal code rule
 - Discover card validation rule
 - Portuguese postal code rule
+- Spanish VAT number
+- Domain name rule
 
 ### Changed