+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Greek Postal code rule
 - Added BSB number rule
 - Added docker-compose for development
+- Added Laravel 9.x support
+- Added support for PHP 8.1
 
 ### Changed
 - Refactored RBG and RGBA rules